fix: Use full month names in Uzbek for attendance dates (Fevral instead of M02)

+++ app/(dashboard)/parent/attendance/parent-attendance-view.tsx
@@ -193,7 +193,7 @@ export function ParentAttendanceView({
   const formatShortDate = (date: Date) => {
     return new Date(date).toLocaleDateString('uz-UZ', {
       day: '2-digit',
-      month: 'short'
+      month: 'long'
     })
   }
